Scheduler bugfixes

+++ src/main/java/com/webs/J15t98J/MaintenanceManager/ScheduleObject.java
@@ -5,13 +5,13 @@ import java.time.LocalDateTime;
 
 public class ScheduleObject {
 
-    public int id;
+    public Long id;
     public LocalDateTime start;
     public Duration duration;
     public boolean kickPlayers;
     public boolean persistRestart;
 
-    public ScheduleObject(LocalDateTime start, Duration duration, boolean kickPlayers, boolean persistRestart, int itemID) {
+    public ScheduleObject(LocalDateTime start, Duration duration, boolean kickPlayers, boolean persistRestart, Long itemID) {
         this.id = itemID;
         this.start = start;
         this.duration = duration;
